+++ test/_env/util/ws.js
@@ -5,11 +5,13 @@ const WebSocket = require("ws");
 
 const auth = require("./auth");
 
-async function connect(service, options = {}) {
+async function connect(service, options = {}, headers = {}, protoocls) {
   const port = cds.app.server.address().port;
-  const socket = new WebSocket(`ws://localhost:${port}` + service, {
+  protoocls ??= [];
+  const socket = new WebSocket(`ws://localhost:${port}` + service, protoocls, {
     headers: {
       authorization: options?.authorization || auth.alice,
+      ...headers,
     },
   });
   cds.wss.once("connection", async (serverSocket) => {
@@ -80,12 +82,13 @@ async function waitForNoEvent(socket, event, timeout = 100) {
   });
 }
 
-async function waitForMessage(socket, event, cb) {
+async function waitForMessage(socket, event, cb, parse) {
   _initListeners(socket);
   return new Promise((resolve) => {
     socket._listeners.push((message) => {
       message = message.toString();
       if (message.includes(event)) {
+        message = parse ? JSON.parse(message) : message;
         resolve(message);
         cb && cb(message);
       }