+++ test/ws/redis_ws.test.js
@@ -52,11 +52,12 @@ describe("Redis", () => {
     expect(redis.client.connect).toHaveBeenCalledWith();
     expect(redis.client.on).toHaveBeenNthCalledWith(1, "error", expect.any(Function));
     expect(redis.client.subscribe).toHaveBeenNthCalledWith(1, "websocket/chat", expect.any(Function));
-    expect(redis.client.subscribe).toHaveBeenNthCalledWith(2, "websocket/fns-websocket", expect.any(Function));
-    expect(redis.client.subscribe).toHaveBeenNthCalledWith(3, "websocket/main", expect.any(Function));
+    expect(redis.client.subscribe).toHaveBeenNthCalledWith(2, "websocket/cloudevent", expect.any(Function));
+    expect(redis.client.subscribe).toHaveBeenNthCalledWith(3, "websocket/fns-websocket", expect.any(Function));
+    expect(redis.client.subscribe).toHaveBeenNthCalledWith(4, "websocket/main", expect.any(Function));
     expect(redis.client.publish).toHaveBeenCalledWith(
       "websocket/chat",
-      `{"event":"received","data":{"text":"test","user":"alice"},"tenant":"t1"}`,
+      '{"event":"received","data":{"text":"test","user":"alice"},"tenant":"t1","headers":{"header":"value"}}',
     );
 
     // Duplicated because Redis mock publishes to same client (not done for real Redis)